test(PurchaseParserTest): 구매 입력 예외 테스트 코드 작성
상품명에 알파벳과 한글 이외의 문자가 들어올 경우의 예외 테스트 코드 작성
