+++ README.md
@@ -19,8 +19,12 @@
     - [X] 안내 문구는 “`구매하실 상품명과 수량을 입력해 주세요. (예: [사이다-2],[감자칩-1])`”이다.
 - [X] 구매 상품과 수량을 입력받는다.
     - [X] 상품명, 수량은 하이픈(-)으로, 개별 상품은 대괄호([])로 묶어 쉼표(,)로 구분한다.
+    - [ ] 상품명에 알파벳과 한글 이외의 입력이 들어올 경우 예외가 발생한다.
+    - [ ] 수량이 음의 정수일 경우 예외가 발생한다.
+    - [ ] 수량이 1000개를 초과할 경우 예외가 발새앟ㄴ다.
     - [ ] 입력 형식 외의 형식이 입력될 경우 예외가 발생한다.
     - [ ] 입력 형식에 포함되지 않는 문자가 입력될 경우 예외가 발생한다.
+    - [ ] 입력 형식에 포함된 문자가 입력되지 않을 경우 예외가 발생한다.
     - [ ] 아무것도 입력되지 않을 경우(””) 예외가 발생한다.
     - [ ] null이 입력될 경우 예외가 발생한다.
     - [ ] 입력한 상품이 없는 상품이 없을 경우 예외가 발생한다.
